Fixed pipe and other things

+++ Kernel/Devices/Keyboard/KeyCodes.d
@@ -1,7 +1,7 @@
 module Devices.Keyboard.KeyCodes;
 
 
-enum Key : short {
+enum Key : byte {
 	Null = -1,
 
 	A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z,
